Renamed View into desktopView and extracted common ancestor with LayerSurface to View
Also did some ligh cleanup
