+++ include/Output.hpp
@@ -21,9 +21,9 @@ public:
   ~Output() noexcept;
 
   void setFrameListener();
-  void setFullscreenView(View *view) noexcept;
+  void setFullscreenView(XdgView *view) noexcept;
 
-  View *getFullscreenView() const noexcept
+  XdgView *getFullscreenView() const noexcept
   {
     return fullscreenView;
   }
@@ -40,7 +40,7 @@ public:
 private:
   std::vector<std::unique_ptr<Workspace>> workspaces;
   struct wlr_output *wlr_output;
-  View *fullscreenView;
+  XdgView *fullscreenView;
   wlr_texture *wallpaperTexture;
   std::array<std::vector<std::unique_ptr<LayerSurface>>, 4> layers;
