+++ include/OutputManager.hpp
@@ -7,6 +7,8 @@
 #include <vector>
 #include <memory>
 
+class View;
+
 /* Used to move all of the data necessary to render a surface from the top-level
  * frame handler to the per-surface render function. */
 struct render_data
@@ -19,17 +21,6 @@ struct render_data
 };
 
 class LayerSurface;
-
-/* Used to move all of the data necessary to render a surface from the top-level
- * frame handler to the per-surface render function. */
-struct layer_render_data
-{
-  wlr_output *output;
-  wlr_renderer *renderer;
-  LayerSurface *layer_surface;
-  timespec *when;
-  bool fullscreen;
-};
 
 struct OutputManagerListeners
 {
@@ -47,7 +38,6 @@ public:
   void output_frame(wl_listener *listener, void *data);
   void server_new_output(wl_listener *listener, void *data);
   static void render_surface(wlr_surface *surface, int sx, int sy, void *data);
-  static void render_layer_surface(wlr_surface *surface, int sx, int sy, void *data);
 
   wlr_output_layout *getLayout() const noexcept;
   std::vector<std::unique_ptr<Output>> const& getOutputs() const;
