+++ include/Workspace.hpp
@@ -20,7 +20,7 @@ public:
     return windowTree;
   }
 
-  std::vector<std::unique_ptr<View>> &getViews()
+  std::vector<std::unique_ptr<XdgView>> &getViews()
   {
     return views;
   };
@@ -31,6 +31,6 @@ public:
 private:
   Output &output;
 
-  std::vector<std::unique_ptr<View>> views;
+  std::vector<std::unique_ptr<XdgView>> views;
   wm::WindowTree windowTree;
 };
