+++ source/Keyboard.cpp
@@ -1,11 +1,12 @@
 #include <unistd.h>
 #include <iostream>
 #include <algorithm>
+#include <memory>
 
 #include "Keyboard.hpp"
 #include "Commands.hpp"
 #include "Server.hpp"
-
+#include "XdgView.hpp"
 
 std::map<std::string, uint32_t> modifiersLst = {
   {"Alt", WLR_MODIFIER_ALT},
