+++ source/XdgShell.cpp
@@ -1,6 +1,7 @@
 #include <cassert>
 
 #include "XdgShell.hpp"
+#include "XdgView.hpp"
 #include "Server.hpp"
 #include "wm/Container.hpp"
 
@@ -13,20 +14,20 @@ XdgShell::XdgShell() {
 void XdgShell::xdg_surface_destroy(wl_listener *listener, void *data)
 {
   Server &server = Server::getInstance();
-  View *view = wl_container_of(listener, view, destroy);
+  XdgView *view = wl_container_of(listener, view, destroy);
 
   if (server.getViews().front().get() == view)
     {
       server.seat.getSeat()->keyboard_state.focused_surface = nullptr;
     }
   server.getViews().erase(std::find_if(server.getViews().begin(), server.getViews().end(),
-				       [view](auto const &ptr)
+				       [view](auto const &ptr) noexcept
 				       {
 					 return ptr.get() == view;
 				       }));
   if (!server.getViews().empty())
     {
-      std::unique_ptr<View> &currentView = server.getViews().front();
+      auto &currentView = server.getViews().front();
       currentView->focus_view();
     }
 };
@@ -36,5 +37,5 @@ void XdgShell::server_new_xdg_surface(wl_listener *listener, void *data)
   wlr_xdg_surface *xdg_surface = static_cast<wlr_xdg_surface *>(data);
 
   if (xdg_surface->role == WLR_XDG_SURFACE_ROLE_TOPLEVEL)
-    Server::getInstance().getViews().emplace_back(new View(xdg_surface->surface));
+    Server::getInstance().getViews().emplace_back(new XdgView(xdg_surface->surface));
 };
